src/services/agent/pyproject.toml
[project]
name = "eletromidia-agent"
version = "0.1.0"
description = "OOH media planning AI agent"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.95.0",
  "uvicorn[standard]>=0.21.0",
  "python-dotenv>=1.0.0",
  "google-cloud-bigquery>=3.20.0",
  "langchain-groq",
  "langchain-google-genai",
  "langchain-core>=0.3.0",
  "langgraph>=0.2.0",
  "httpx>=0.27.0",
  "numpy",
  "pandas",
  "pyarrow>=23.0.1",
]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["UP045"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
